requirements.txt
litellm==1.74.3
google-adk==1.6.1
openai==1.95.1
litellm==1.74.3
python-dotenv==1.1.1
